DPRO-3346: Move ThemeDebugInfo out from inner class
The graph-traversal code comprises enough clutter to be worth removing from
the ThemeTree class.
